requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
requests==2.32.3
APScheduler==3.10.4
Jinja2==3.1.6
python-dotenv==1.0.1
geopy==2.4.1
playwright==1.52.0
